pyproject.toml
[project]
name = "recruitpro"
version = "0.1.0"
description = "RecruitPro ATS and AI platform"
authors = [{ name = "RecruitPro Team", email = "dev@recruitpro.local" }]
requires-python = ">=3.11,<4.0"
dependencies = [
  "fastapi>=0.111.0,<0.112.0",
  "uvicorn[standard]>=0.30.0,<0.31.0",
  "sqlalchemy>=2.0.30,<3.0.0",
  "alembic>=1.13.1,<2.0.0",
  "passlib[bcrypt]>=1.7.4,<2.0.0",
  "python-jose>=3.3.0,<4.0.0",
  "python-multipart>=0.0.9,<0.1.0",
  "pydantic-settings>=2.3.0,<3.0.0",
  "email-validator>=2.1.1,<3.0.0",
  "openpyxl>=3.1.2,<4.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.1,<9.0.0",
  "httpx>=0.27.0,<0.28.0",
  "anyio>=4.3.0,<5.0.0",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"
